[iron] Rename algorithms/transform.py to _transform.py
Fixes a Python name collision: `aie.iron.algorithms.transform` resolved
to either the public callable or the submodule depending on import order.
Renaming the module to `_transform` (private convention) leaves
`algorithms.transform` unambiguously the callable.
